implement randomkey no examples
