+++ java/integTest/src/test/java/glide/cluster/CommandTests.java
@@ -40,6 +40,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assumptions.assumeTrue;
 
+import glide.api.BaseClient;
 import glide.api.RedisClusterClient;
 import glide.api.models.ClusterTransaction;
 import glide.api.models.ClusterValue;
@@ -1544,5 +1545,24 @@ public class CommandTests {
                 checkFunctionStatsResponse(nodeResponse, new String[0], 0, 0);
             }
         }
+    }
+
+    @Test
+    @SneakyThrows
+    public void randomKey() {
+        String key1 = "{key}" + UUID.randomUUID();
+        String key2 = "{key}" + UUID.randomUUID();
+
+        assertEquals(OK, clusterClient.set(key1, "a").get());
+        assertEquals(OK, clusterClient.set(key2, "b").get());
+
+        String randomKey = clusterClient.randomKey().get();
+        assertEquals(1L, clusterClient.exists(new String[] {randomKey}).get());
+
+        // no keys in database
+        assertEquals(OK, clusterClient.flushall().get());
+        ExecutionException executionException =
+            assertThrows(ExecutionException.class, () -> clusterClient.randomKey().get());
+        assertInstanceOf(RequestException.class, executionException.getCause());
     }
 }
